+++ packages/client/src/rtc/__tests__/videoLayers.test.ts
@@ -3,6 +3,7 @@ import { describe, expect, it, vi } from 'vitest';
 import {
   findOptimalScreenSharingLayers,
   findOptimalVideoLayers,
+  getComputedMaxBitrate,
 } from '../videoLayers';
 
 describe('videoLayers', () => {
@@ -34,7 +35,11 @@ describe('videoLayers', () => {
     const targetBitrate = 3000000;
     vi.spyOn(track, 'getSettings').mockReturnValue({ width, height });
 
-    const layers = findOptimalVideoLayers(track, targetBitrate);
+    const layers = findOptimalVideoLayers(track, {
+      width,
+      height,
+      bitrate: targetBitrate,
+    });
     expect(layers).toEqual([
       {
         active: true,
@@ -97,5 +102,37 @@ describe('videoLayers', () => {
     expect(layers[0].rid).toBe('q');
     expect(layers[1].rid).toBe('h');
     expect(layers[2].rid).toBe('f');
+  });
+
+  describe('getComputedMaxBitrate', () => {
+    it('should scale target bitrate down if resolution is smaller than target resolution', () => {
+      const targetResolution = { width: 1920, height: 1080, bitrate: 3000000 };
+      const scaledBitrate = getComputedMaxBitrate(targetResolution, 1280, 720);
+      expect(scaledBitrate).toBe(1333333);
+    });
+
+    it('should not scale target bitrate if resolution is larger than target resolution', () => {
+      const targetResolution = { width: 1280, height: 720, bitrate: 1000000 };
+      const scaledBitrate = getComputedMaxBitrate(targetResolution, 2560, 1440);
+      expect(scaledBitrate).toBe(1000000);
+    });
+
+    it('should not scale target bitrate if resolution is equal to target resolution', () => {
+      const targetResolution = { width: 1280, height: 720, bitrate: 1000000 };
+      const scaledBitrate = getComputedMaxBitrate(targetResolution, 1280, 720);
+      expect(scaledBitrate).toBe(1000000);
+    });
+
+    it('should handle 0 width and height', () => {
+      const targetResolution = { width: 1280, height: 720, bitrate: 1000000 };
+      const scaledBitrate = getComputedMaxBitrate(targetResolution, 0, 0);
+      expect(scaledBitrate).toBe(0);
+    });
+
+    it('should handle 4k target resolution', () => {
+      const targetResolution = { width: 3840, height: 2160, bitrate: 15000000 };
+      const scaledBitrate = getComputedMaxBitrate(targetResolution, 1280, 720);
+      expect(scaledBitrate).toBe(1666667);
+    });
   });
 });