+++ packages/react-bindings/src/hooks/permissions.ts
@@ -1,5 +1,7 @@
-import { OwnCapability } from '@stream-io/video-client';
+import { OwnCapability, PermissionRequestEvent } from '@stream-io/video-client';
 import { useCallMetadata } from './call';
+import { useCallState } from './store';
+import { useObservableValue } from './helpers/useObservableValue';
 
 /**
  * Hook that returns true if the current user has all the given permissions.
@@ -24,4 +26,16 @@ export const useHasPermissions = (...permissions: OwnCapability[]) => {
 export const useOwnCapabilities = (): OwnCapability[] => {
   const metadata = useCallMetadata();
   return metadata?.own_capabilities || [];
+};
+
+/**
+ * A hook which returns the latest call permission request.
+ *
+ * @category Call State
+ */
+export const useCallPermissionRequest = ():
+  | PermissionRequestEvent
+  | undefined => {
+  const { callPermissionRequest$ } = useCallState();
+  return useObservableValue(callPermissionRequest$);
 };