+++ packages/react-sdk/src/components/Debug/DebugStatsView.tsx
@@ -1,15 +1,26 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Call } from '@stream-io/video-client';
-import { useRtcStats } from '../../hooks/useRtcStats';
 import { useFloatingUIPreset } from '../../hooks';
+import { StatCard } from '../CallStats';
+import { useCallStatsReport } from '@stream-io/video-react-bindings';
 
 export const DebugStatsView = (props: {
   call: Call;
-  kind: 'subscriber' | 'publisher';
   mediaStream?: MediaStream;
+  sessionId: string;
 }) => {
-  const { call, kind, mediaStream } = props;
-  const stats = useRtcStats(call, kind, mediaStream);
+  const { call, mediaStream, sessionId } = props;
+  const callStatsReport = useCallStatsReport();
+
+  useEffect(() => {
+    call.startReportingStatsFor(sessionId);
+    return () => {
+      call.stopReportingStatsFor(sessionId);
+    };
+  }, [call, sessionId]);
+
+  const reportForTracks = callStatsReport?.participants[sessionId];
+  const trackStats = reportForTracks?.flatMap((report) => report.streams);
 
   const { refs, strategy, y, x } = useFloatingUIPreset({
     placement: 'top',
@@ -38,7 +49,7 @@ export const DebugStatsView = (props: {
       />
       {isPopperOpen && (
         <div
-          className="str-video__debug__track-stats"
+          className="str-video__debug__track-stats str-video__call-stats"
           ref={refs.setFloating}
           style={{
             position: strategy,
@@ -47,9 +58,51 @@ export const DebugStatsView = (props: {
             overflowY: 'auto',
           }}
         >
-          <pre>{JSON.stringify(stats, null, 2)}</pre>
+          <h3>Participant stats</h3>
+          <div className="str-video__call-stats__card-container">
+            {trackStats
+              ?.map((track) => {
+                if (track.kind === 'video') {
+                  return (
+                    <StatCard
+                      key={`${track.rid}/${track.ssrc}/${track.codec}/${track.kind}`}
+                      label={
+                        `${track.kind}: ${track.codec} ` +
+                        (track.rid ? ` (${track.rid})` : '')
+                      }
+                      value={`${track.frameWidth || 0}x${
+                        track.frameHeight || 0
+                      }@${track.framesPerSecond || 0}fps`}
+                    />
+                  );
+                } else if (track.kind === 'audio') {
+                  return (
+                    <StatCard
+                      key={`${track.ssrc}/${track.codec}/${track.kind}`}
+                      label={track.codec || 'N/A'}
+                      value={`Jitter: ${track.jitter || 0}ms`}
+                    />
+                  );
+                }
+                return null;
+              })
+              .filter(Boolean)}
+          </div>
+          {reportForTracks?.map((report, index) => (
+            <pre key={index}>
+              {JSON.stringify(unwrapStats(report.rawStats), null, 2)}
+            </pre>
+          ))}
         </div>
       )}
     </>
   );
+};
+
+const unwrapStats = (rawStats?: RTCStatsReport) => {
+  const decodedStats: Record<string, string> = {};
+  rawStats?.forEach((s) => {
+    decodedStats[s.id] = s;
+  });
+  return decodedStats;
 };